Move tests for functions now in `db-config`
